requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Data generation and CLI
faker>=18.0.0
click>=8.0.0
python-dateutil>=2.8.0

# Testing and development
pytest>=7.0.0
pytest-cov>=4.0.0

# Configuration management
pyyaml>=6.0

# Optional: Enhanced performance
# numba>=0.56.0

# Documentation (optional)
# sphinx>=5.0.0
# sphinx-rtd-theme>=1.0.0
